reorganize project files
