+++ app/src/main/java/com/personalfinancetracer/Main.kt
@@ -1,4 +1,4 @@
-import finance.app.FinanceApp
+import com.personalfinancetracer.finance.app.FinanceApp
 
 fun main() {
     val financeApp = FinanceApp()
