+++ app/src/main/java/com/personalfinancetracer/finance/app/Command.kt
@@ -1,0 +1,6 @@
+package com.personalfinancetracer.finance.app
+
+interface Command {
+     fun execute()
+}
+
