+++ app/src/main/java/com/personalfinancetracer/finance/app/usecase/GenerateCategorySummaryUseCase.kt
@@ -1,0 +1,9 @@
+package com.personalfinancetracer.finance.app.usecase
+
+import com.personalfinancetracer.finance.app.Command
+
+class GenerateCategorySummaryUseCase: Command{
+    override fun execute() {
+        TODO("Not yet implemented")
+    }
+}
