+++ app/src/main/java/com/personalfinancetracer/finance/app/usecase/GenerateMonthlySummaryUseCase.kt
@@ -1,0 +1,9 @@
+package com.personalfinancetracer.finance.app.usecase
+
+import com.personalfinancetracer.finance.app.Command
+
+class GenerateMonthlySummaryUseCase:Command {
+    override fun execute() {
+        TODO("Not yet implemented")
+    }
+}
