+++ app/src/main/java/com/personalfinancetracer/finance/app/usecase/GenerateReportUseCase.kt
@@ -1,0 +1,10 @@
+package com.personalfinancetracer.finance.app.usecase
+
+import com.personalfinancetracer.finance.app.Command
+
+class GenerateReportUseCase: Command {
+    override fun execute() {
+        TODO("Not yet implemented")
+    }
+
+}
