+++ app/src/main/java/com/personalfinancetracer/finance/app/usecase/ListAllTransactionsUseCase.kt
@@ -1,0 +1,11 @@
+package com.personalfinancetracer.finance.app.usecase
+
+import com.personalfinancetracer.finance.app.Command
+
+class ListAllTransactionsUseCase:Command
+{
+    override fun execute() {
+        TODO("Not yet implemented")
+    }
+
+}
